Also peform Xcode build
Previously this script just build the xcodeproj from bazel. This was
helpful, however sometimes I want to debug with Xcode, which requires
running from Xcode. Before, I had to wait for the bazel build to finish
before launching the Xcode build. Now, this script does both so I can
fire and foreget (and work on other tasks in the meantime)